Compatible with non-generic versions of Go
