+++ export_test.go
@@ -9,8 +9,7 @@ const (
 )
 
 type (
-	BucketPadded   = bucketPadded
-	BucketOfPadded = bucketOfPadded
+	BucketPadded = bucketPadded
 )
 
 type MapStats struct {
@@ -63,19 +62,4 @@ func MakeSeed() uint64 {
 
 func HashString(s string, seed uint64) uint64 {
 	return hashString(s, seed)
-}
-
-func MakeHasher[T comparable]() func(T, uint64) uint64 {
-	return makeHasher[T]()
-}
-
-func CollectMapOfStats[K comparable, V any](m *MapOf[K, V]) MapStats {
-	return MapStats{m.stats()}
-}
-
-func NewMapOfPresizedWithHasher[K comparable, V any](
-	hasher func(K, uint64) uint64,
-	sizeHint int,
-) *MapOf[K, V] {
-	return newMapOfPresized[K, V](hasher, sizeHint)
 }
